+++ lib/pages/account_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import './login_page.dart';
+import 'login_page.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 
 class Account extends StatelessWidget {
